requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-genai==1.13.0
pdfplumber==0.10.2